+++ ConferenceApp/Controllers/TalkController.cs
@@ -42,7 +42,7 @@ namespace ConferenceApp.Controllers
                 return NotFound();
             }
             var currentUserId = HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
-            var isAssistant = await _context.Roles.Where(x => (x.UserId == currentUserId && x.EventId == talk.Id)).ToListAsync();
+            var isAssistant = await _context.Roles.Where(x => (x.UserId == currentUserId && x.EventId == talk.Id && x.Name == "attendant")).ToListAsync();
 
             int assisting = isAssistant.Count;
             ViewBag.assisting = assisting;
@@ -74,14 +74,14 @@ namespace ConferenceApp.Controllers
                 sponsors.Add(s.Name);
             }
 
-            var assistantRoles = await _context.Roles.Where(x => x.EventId == @talk.Id).ToListAsync();
+            var assistantRoles = await _context.Roles.Where(x => x.EventId == @talk.Id && x.Name == "attendant").ToListAsync();
             var assistants = new List<object>();
             // foreach (var member in assistantRoles)
             // {
             //     var a = await _context.Users.FindAsync(member.UserId);
             //     assistants.Add(a.Email);
             // }
-            
+
             var fileDescription = "";
             if (talk.FileId > 0)
             {
@@ -91,6 +91,50 @@ namespace ConferenceApp.Controllers
                     fileDescription = file.Description;
                 }
             }
+            
+            var exhibitor = await _context.Users.FindAsync(@talk.Exhibitor);
+            
+            ViewBag.Exhibitor = exhibitor;
+
+            var EventAssistance = await _context.Roles.Where(x => x.EventId == talk.Id && x.Name == "attendant").CountAsync();
+
+            var FeedbackCategories = await _context.FeedbackCategories.ToListAsync();
+            var Feedbacks = await _context.Feedbacks.Where(x => x.EventId == talk.Id).ToListAsync();
+
+            var FeedbackAveragePerCategory = new List<object>();
+            var FeedbackCategoryName = new List<object>();
+
+            foreach (var Category in FeedbackCategories)
+            {
+                FeedbackCategoryName.Add(Category.Name);
+                var FeedbacksScopesOfEventAndCategory = await _context.FeedbackScopes.Where(fs => Feedbacks.Any(f => fs.FeedbackId == f.Id && fs.FeedbackCategoryId == Category.Id)).ToListAsync();
+
+                if (FeedbacksScopesOfEventAndCategory.Count() >= 1)
+                {
+                    FeedbackAveragePerCategory.Add(FeedbacksScopesOfEventAndCategory.Average(f => f.Grade).ToString());
+                }
+                else
+                {
+                    FeedbackAveragePerCategory.Add("No hay evaluaciones todavía");
+                }
+            }
+            ViewBag.permision = false;
+
+            var conferenceVersion = await _context.ConferenceVersions.FindAsync(@talk.ConferenceVersionId);
+            var bigConference = await _context.Conferences.FindAsync(conferenceVersion.ConferenceId);
+            var adminList = await _context.Admins.Where(x => x.UserId == currentUserId).ToListAsync();
+            if (adminList.Count > 0)
+            {
+                ViewBag.permision = true;
+            }
+
+            if (bigConference.OrganizerId == currentUserId)
+            {
+                ViewBag.permision = true;
+            }
+            var organizer = await _context.Users.FindAsync(bigConference.OrganizerId);
+            ViewBag.organizer = organizer;
+            ViewBag.feedback = await _context.Feedbacks.FirstOrDefaultAsync(f => f.UserId == currentUserId && f.EventId == @talk.Id);
 
             ViewBag.roomName = room.Name;
             ViewBag.centreName = centre.Name;
@@ -100,6 +144,16 @@ namespace ConferenceApp.Controllers
             ViewBag.assistants = assistants;
             ViewBag.sponsors = sponsors;
             ViewBag.fileDescription = fileDescription;
+            ViewBag.EventAssistance = EventAssistance;
+            ViewBag.FeedbackCategoryName = FeedbackCategoryName;
+            ViewBag.FeedbackAveragePerCategory = FeedbackAveragePerCategory;
+
+            var isOrganizer = currentUserId == conference.OrganizerId;
+            ViewBag.isOrganizer = isOrganizer;
+            
+            var admin = await _context.Admins.FirstOrDefaultAsync(x => x.UserId == currentUserId);
+            var isAdmin = admin != null;
+            ViewBag.isAdmin = isAdmin;
 
             return View(talk);
         }
@@ -122,7 +176,14 @@ namespace ConferenceApp.Controllers
 
             var tags = await _context.Tags.ToListAsync();
             var availableTags = tags.Select(tag => new CheckBoxItem() {TagId = tag.Id, Title = tag.Name, IsChecked = false}).ToList();
-
+            
+            var users = await _context.Users.ToListAsync();
+            var userList = new List<object>();
+            foreach (var user in users)
+            {
+                userList.Add(user);
+            }
+            ViewBag.Exhibitors = new SelectList(userList , "Id", "Email");
             this.ViewData["ConferenceVersions"] = new SelectList(versions, "Id", "Name");
             this.ViewData["Rooms"] = new SelectList(rooms, "Id", "Name");
             this.ViewData["AvailableTags"] = new List<CheckBoxItem>(availableTags);
@@ -134,7 +195,7 @@ namespace ConferenceApp.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Topic,Id,Name,StartDate,EndDate,ConferenceVersionId,RoomId,Exhibitor,AvailableTags")] Talk talk)
+        public async Task<IActionResult> Create([Bind("Topic,Id,Name,StartDate,EndDate,ConferenceVersionId,RoomId,ExhibitorId,AvailableTags")] Talk talk)
         {
 
             var conferenceVersion = await _context.ConferenceVersions.Where(x => x.Id == talk.ConferenceVersionId).FirstOrDefaultAsync();
@@ -179,14 +240,17 @@ namespace ConferenceApp.Controllers
                     if (ModelState.IsValid)
                     {
                         var eventTags = new List<EventTag>();
-                        for (var i = 0; i < talk.AvailableTags.Count; i++)
+                        if (talk.AvailableTags != null)
                         {
-                            if (talk.AvailableTags[i].IsChecked)
+                            for (var i = 0; i < talk.AvailableTags.Count; i++)
                             {
-                                var tag = await _context.Tags.FirstOrDefaultAsync(m => m.Id == talk.AvailableTags[i].TagId);
-                                var eventTag = new EventTag() {Event = talk, EventId = talk.Id, Tag = tag, TagId = tag.Id};
-                                _context.Add(eventTag);
-                                eventTags.Add(eventTag);
+                                if (talk.AvailableTags[i].IsChecked)
+                                {
+                                    var tag = await _context.Tags.FirstOrDefaultAsync(m => m.Id == talk.AvailableTags[i].TagId);
+                                    var eventTag = new EventTag() {Event = talk, EventId = talk.Id, Tag = tag, TagId = tag.Id};
+                                    _context.Add(eventTag);
+                                    eventTags.Add(eventTag);
+                                }
                             }
                         }
                         talk.EventTags = eventTags;
@@ -224,6 +288,13 @@ namespace ConferenceApp.Controllers
                 var checkBox = new CheckBoxItem() {TagId = tag.Id, Title = tag.Name, IsChecked = eventTag != null};
                 availableTags.Add(checkBox);
             }
+            var users = await _context.Users.ToListAsync();
+            var userList = new List<object>();
+            foreach (var user in users)
+            {
+                userList.Add(user);
+            }
+            ViewBag.Exhibitors = new SelectList(userList , "Id", "Email");
             this.ViewData["AvailableTags"] = availableTags;
 
             return View(talk);
@@ -233,7 +304,7 @@ namespace ConferenceApp.Controllers
         {
 
             var currentUserId = HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
-            var assistants = await _context.Roles.Where(x => (x.UserId == currentUserId && x.EventId == eventId)).ToListAsync();
+            var assistants = await _context.Roles.Where(x => (x.UserId == currentUserId && x.EventId == eventId && x.Name == "attendant")).ToListAsync();
             _context.Roles.RemoveRange(assistants);
             await _context.SaveChangesAsync();
 
@@ -245,7 +316,7 @@ namespace ConferenceApp.Controllers
             var currentUserId = HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
             var @thisEvent = await _context.Events.FirstOrDefaultAsync(m => m.Id == eventId);
             var isOccupied = 0;
-            
+
             var room = await _context.Rooms.FirstOrDefaultAsync(m => m.Id == @thisEvent.RoomId);
             var capacityUsed = await _context.Roles.Where(x => (x.EventId == eventId && x.Name == "attendant")).ToListAsync();
 
@@ -256,7 +327,7 @@ namespace ConferenceApp.Controllers
             }
             else
             {
-                var assistingToEvents = await _context.Roles.Where(x => (x.UserId == currentUserId)).ToListAsync();
+                var assistingToEvents = await _context.Roles.Where(x => (x.UserId == currentUserId && x.Name == "attendant")).ToListAsync();
                 foreach (var aRole in assistingToEvents)
                 {
                     var @event = await _context.Events.FirstOrDefaultAsync(m => m.Id == aRole.EventId);
@@ -283,7 +354,7 @@ namespace ConferenceApp.Controllers
                     }
                 }
             }
-            
+
             if (isOccupied == 0)
             {
                 var role = new Role() {UserId = currentUserId, EventId = eventId};
@@ -293,7 +364,6 @@ namespace ConferenceApp.Controllers
             return RedirectToAction(nameof(Details), new { id = eventId.ToString() });
 
         }
-
 
         // POST: Talk/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for
@@ -312,26 +382,29 @@ namespace ConferenceApp.Controllers
                 try
                 {
                     var eventTags = new List<EventTag>();
-                    for (var i = 0; i < talk.AvailableTags.Count; i++)
+                    if (talk.AvailableTags != null)
                     {
-                        var existingEventTag = await _context.EventTags.FirstOrDefaultAsync(x => x.TagId == talk.AvailableTags[i].TagId && x.EventId == talk.Id);
-                        if (talk.AvailableTags[i].IsChecked)
+                        for (var i = 0; i < talk.AvailableTags.Count; i++)
                         {
-                            // si el tag está checkeado y ya exitía este eventTag, no hacer nada, sino crearlo
-                            if (existingEventTag == null)
+                            var existingEventTag = await _context.EventTags.FirstOrDefaultAsync(x => x.TagId == talk.AvailableTags[i].TagId && x.EventId == talk.Id);
+                            if (talk.AvailableTags[i].IsChecked)
                             {
-                                var tag = await _context.Tags.FirstOrDefaultAsync(x => x.Id == talk.AvailableTags[i].TagId);
-                                var newEventTag = new EventTag() {Event = talk, EventId = talk.Id, Tag = tag, TagId = tag.Id};
-                                _context.Add(newEventTag);
-                                eventTags.Add(newEventTag);
+                                // si el tag está checkeado y ya exitía este eventTag, no hacer nada, sino crearlo
+                                if (existingEventTag == null)
+                                {
+                                    var tag = await _context.Tags.FirstOrDefaultAsync(x => x.Id == talk.AvailableTags[i].TagId);
+                                    var newEventTag = new EventTag() {Event = talk, EventId = talk.Id, Tag = tag, TagId = tag.Id};
+                                    _context.Add(newEventTag);
+                                    eventTags.Add(newEventTag);
+                                }
                             }
-                        }
-                        else
-                        {
-                            // si no está checkeado y ya exitía este eventTag, eliminarlo, sino no hacer nada
-                            if (existingEventTag != null)
+                            else
                             {
-                                _context.EventTags.Remove(existingEventTag);
+                                // si no está checkeado y ya exitía este eventTag, eliminarlo, sino no hacer nada
+                                if (existingEventTag != null)
+                                {
+                                    _context.EventTags.Remove(existingEventTag);
+                                }
                             }
                         }
                     }
